request.py ----------
add: new property: 'app'. This property contains an instance of Sanic application.
add: new property: 'handler'. This property will be setted before a request is sent to the view or middleware and contains the handler of view. This is useful to know all information of view, for example: if any plugin decorates the function, only in runtime the middleware or view will know this information. The most common case is when we want to decorate a view o use a cache system. The middleware must know who view are marked as a cached.

sanic.py
--------

add: new method to append a middleware. The idea is to set middleware easily, thinking in possible plugins. Using this way, a plugin could use this method to add their custom middleware. For example: An authentication system or cache.
modified: reordered the the call of "self.router.get(...)", to add the "handler" property to the request before any middleware is called.

server.py
---------

add: app instance in constructors