Removes the need for asycnio global settings
Also moves sanic.py -> app.py and removes the need for us to relative
import anymore.

This is a step towards allowing people to use things like
`sanic.get_event_loop()` instead of having to set global asyncio
settings.

Does not seem to work on 3.5.1 but works fine on every version after
that.