scheme arg to create the redirect should be 'https'
Since this example is trying to redirect HTTP requests to the HTTPS server, the scheme should be 'https.' I haven't tried running the example by itself, but I'm surprised it works as it is. Perhaps it is because of https.config.SERVER_NAME, but I wouldn't expect that since it doesn't include the scheme (but does include the port).